Uploading Manually
Files failed to upload from eclipse.  Files changed are setup and tictactoe.  Minimax was created to hold the methods required for the minimax algorithm.  Optimized setup for better readability and code reduction.  Changed linked list in tictactoe class to a hashmap to optimize the inserts and removals in minimax alg.